requirements.txt
flet==0.28.3
flask==3.1.2
faster-whisper==1.2.0
huggingface-hub==0.34.4
opencv-python==4.12.0.88
mediapipe==0.10.14
piper-tts==1.3.0
PyAudio==0.2.14
aiohttp==3.12.15
webrtcvad==2.0.10
PyMuPDF==1.26.4
